+++ game_patch/rf/v3d.h
@@ -144,7 +144,7 @@ namespace rf
     {
         MRF_SCANNER_1 = 1,
         MRF_SCANNER_2 = 8,
-        MRF_AMBIENT_COLOR = 0x80,
+        MRF_CUSTOM_AMBIENT_COLOR = 0x80,
         MRF_CLIP_VERTICES = 0x100,
         MRF_FIRST_PERSON = 0x400,
     };