+++ docs/CHANGELOG.md
@@ -37,6 +37,7 @@ Version 1.2.1 (TBD): Not yet released
 - Fix players spawning at the wrong control point in ESC when the center point is neutral
 - Fix rare crash when shooting at alpha-masked surfaces (fixes crash in community level `dm-oranmentsb2.rfl`)
 - Fix skybox rotation not being rendered by Direct3D 11 renderer
+- Fix third person weapon models being rendered fullbright by Direct3D 11 renderer
 
 [@is-this-c](https://github.com/is-this-c)
 - Use packet queues for remote server config transfers